Update starting1.cs
The current inplementation of the "Controlling spacing"-example is bad practice.
 
Have a look at this video, it explains why it is bad practice:
https://www.youtube.com/watch?v=4lFAs6FYTXg

Conclusion: In for-loops always use StringBuilder instead of string-concatenation. Strings are immutable and as a result you would create many unnecessary objects on the heap by using string-concatenation instead of StringBuilder.